Remove 'UiData'
